Solution to Count the Number of Prime Numbers in an Array

+++ Count the Number of Prime Numbers in an Array@@676c02002dbc8c3e7a280b7d/code.c
@@ -5,7 +5,7 @@ int isPrime(int n) {
     if (n <= 1) {
         return 0;
     }
-    for (int i = 2; i <= sqrt(n); i++) {
+    for (int i = 2; i <= (int)sqrt(n); i++) {
         if (n % i == 0) {
             return 0;
         }
